Created case study in pandas foundations with examples
Added examples of reading file to df, converting string to list and
assigning it as column labels, dropping columns from df, converting date
and time columns to string to concatenate and set as df index. Added
an initial dataset and a filtered dataset generated to load in the
following exercises.