Added Groupby and Transformation to Manipulating DataFrames ch4
Added example of using .transform() method after grouping to apply a
function to groups of data indepenently and filter outliers with z-score
value.
Added example of using .groupby() and .transform() to fill missing data
for each group.
Added example of using .apply() method on groupby object when performs
an arbitrary function on each of the groups.

Added chapter slides